requirements.txt
google-cloud-bigquery>=3.10.0
google-cloud-storage>=2.14.0
mysql-connector-python>=8.3.0
PyYAML>=6.0.1
python-dotenv>=1.0.1
APScheduler>=3.10.4
tenacity>=8.2.3
pytz>=2024.1
pandas>=2.2.2
